Rename eks-a folder to eksctl-anywhere
