+++ cmd/eksctl-anywhere/cmd/delete.go
