+++ cmd/eksctl-anywhere/cmd/generate.go
